Move Patternfly import files to seperate directory
This move will easy copying the directory alone to external projects.

Closes #14502
