+++ pkg/lib/page.scss
@@ -1,4 +1,4 @@
-@import "patternfly-4-overrides.scss";
+@import "./patternfly/patternfly-4-overrides.scss";
 
 a {
     cursor: pointer;
