+++ pkg/lib/patternfly/patternfly-4-cockpit.scss
