+++ pkg/networkmanager/interfaces.js
@@ -17,7 +17,7 @@
  * along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
  */
 
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 import $ from 'jquery';
 import React from "react";
 import ReactDOM from "react-dom";
