+++ pkg/playground/exception.js
@@ -1,5 +1,5 @@
 /* An unhandled javascript exception */
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 import cockpit from "cockpit";
 
 var button = document.getElementById("exception");
