+++ pkg/playground/journal.jsx
@@ -20,7 +20,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 import "page.scss";
 
 import { LogsPanel } from "cockpit-components-logs-panel.jsx";
