+++ pkg/playground/pkgs.js
@@ -1,7 +1,7 @@
 import $ from "jquery";
 import cockpit from "cockpit";
 
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 
 $(function() {
     var proxy = cockpit.dbus(null, { bus: "internal" }).proxy("cockpit.Packages", "/packages");
