+++ pkg/playground/plot.js
@@ -1,7 +1,7 @@
 import $ from "jquery";
 import * as plot from "plot.js";
 
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 import "../../node_modules/@patternfly/patternfly/components/Button/button.scss";
 import "plot.css";
