+++ pkg/playground/speed.js
@@ -1,6 +1,6 @@
 import cockpit from "cockpit";
 
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 import "../../node_modules/@patternfly/patternfly/components/Button/button.scss";
 
 var channel = null;
