+++ pkg/playground/translate.js
@@ -2,7 +2,7 @@ import $ from "jquery";
 import cockpit from "cockpit";
 import { mustache } from "mustache";
 
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 
 const _ = cockpit.gettext;
 var C_ = cockpit.gettext;
