+++ pkg/shell/index.js
@@ -17,7 +17,7 @@
  * along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
  */
 
-import '../../src/base1/patternfly-cockpit.scss';
+import '../lib/patternfly/patternfly-cockpit.scss';
 
 import { machines } from "machines";
 import { new_machine_dialog_manager } from "machine-dialogs";
